onefin/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
factory-boy==3.3.0
Faker==19.13.0
idna==3.4
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
uuid==1.30
